+++ DiscordBot/Database/UploadOnlyRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Discord;
 using DiscordBot.Entity;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
 namespace DiscordBot.Database
@@ -18,16 +19,26 @@ namespace DiscordBot.Database
             _db = db;
         }
 
-        public async Task<UploadOnlyEntity?> Get(int id)
+        public async Task<UploadOnlyEntity?> Get(Guid id)
         {
             return await _db.UploadOnlyEntities.FindAsync(id);
         }
 
+        public async Task<List<UploadOnlyEntity>> GetAll()
+        {
+            return await AsyncEnumerable.ToListAsync(_db.UploadOnlyEntities);
+        }
+
+        public async Task<List<UploadOnlyEntity>> GetAllbyServerId(ulong serverid)
+        {
+            return await _db.UploadOnlyEntities.AsQueryable()
+                .Where(x => x.ServerId == serverid).ToListAsync();
+        }
+
         public async Task<UploadOnlyEntity?> GetByChannelId(ulong channelId)
         {
-            return await _db.UploadOnlyEntities.AsAsyncEnumerable()
-                .Where(a=> a.ChannelId == channelId)
-                .FirstOrDefaultAsync();
+            return await _db.UploadOnlyEntities.AsQueryable()
+                .FirstOrDefaultAsync(a => a.ChannelId == channelId);
         }
         public async Task<bool> Add(UploadOnlyEntity entity)
         {
@@ -59,12 +70,16 @@ namespace DiscordBot.Database
             }
         }
 
-        public async Task<bool> Delete(UploadOnlyEntity entity)
+        public async Task<bool> Delete(Guid id)
         {
             try
             {
-                _db.UploadOnlyEntities.Remove(entity);
-                await _db.SaveChangesAsync();
+                var entity = await _db.UploadOnlyEntities.FindAsync(id);
+                if (entity != null)
+                {
+                    _db.UploadOnlyEntities.Remove(entity);
+                    await _db.SaveChangesAsync();
+                }
                 return true;
             }
             catch (Exception e)
